Moved tests, renamed packages
