+++ crcCards/TeamManager.txt
@@ -1,4 +1,4 @@
-Class name: Team.TeamManager
+Class name: team.TeamManager
 Parent class: None
 Subclasses: None
 Layer: Use Case
@@ -9,4 +9,4 @@ Responsibilities:
 * Find a team from all of the teams
 
 Collaborators:
-* Team.Team
+* team.team
