+++ phase0/specification.md
@@ -8,9 +8,9 @@ Users can upload data on sports players and teams either
 manually through the command line, through a file, or by 
 accessing data from the web via an API (possibly). This data 
 includes a variety of data on teams and players, such as the 
-amount of points a team/Player has, or other statistical information about them.
+amount of points a team/player has, or other statistical information about them.
 Users will be able to examine this data in detail through the app. They can compare players 
-and teams with each other, as well as examine each Player/team individually. This specific data can be exported
+and teams with each other, as well as examine each player/team individually. This specific data can be exported
 as a spreadsheet for the user to easily send someone the information or see it in a different format. The application 
 can generate predictions for players and hypothetical teams based on different kinds of models and data, in order to 
 provide the user with a holistic prediction. 
@@ -18,7 +18,7 @@ provide the user with a holistic prediction.
 Additionally, the user of the app can add members to the app; these members are part of the fantasy 
 betting league of the user, who is the manager of said league. The user
 then can make bets on upcoming matches on the behalf of the members. Each
-league_member will have a history, outlining which matches they were able to 
+leagueMember will have a history, outlining which matches they were able to 
 predict correctly. This will allow the league to have a ranking, and members can compete with each other
 using the app. The aforementioned matches will be added by the user of the app.
 The application will extend to multiple sports, such as tennis, hockey, as well as other sports.
