+++ phase0/walkthrough.md
@@ -5,8 +5,8 @@ rankings and other data of the two teams that are playing against each other in 
 the user with the data that will make it easy and intuitive to compare the two teams. The user will also look at the 
 statistics of some players who will be playing in the upcoming match.
 
-After making a decision on who will  win, the user will create a match in the app, add themselves as a league_member of the 
-fantasy league, and bet on the outcome of the match. Additionally, the user will add their friend as a league_member of the 
+After making a decision on who will  win, the user will create a match in the app, add themselves as a leagueMember of the 
+fantasy league, and bet on the outcome of the match. Additionally, the user will add their friend as a leagueMember of the 
 fantasy league and make a prediction on the same match on their behalf. After the match ends, the user will indicate
 the outcome of the match in the app, and the app will determine which members got the predictions right and which
 didn't.
