+++ src/main/java/player/GetStatsComparator.java
@@ -1,4 +1,4 @@
-package Player;
+package player;
 
 import java.util.Comparator;
