+++ src/main/java/player/HockeyPlayer.java
@@ -1,4 +1,4 @@
-package Player;
+package player;
 
 import java.util.HashMap;
 import java.util.Map;
