+++ src/main/java/player/Player.java
@@ -1,4 +1,4 @@
-package Player;
+package player;
 
 import java.util.Objects;
