+++ src/main/java/player/PlayerList.java
@@ -1,4 +1,4 @@
-package Player;
+package player;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
