+++ src/main/java/sportsApp/SportsApp.java
@@ -1,4 +1,4 @@
-package SportsApp;
+package sportsApp;
 
 import java.util.Scanner;
