+++ src/main/java/team/TennisTeam.java
@@ -1,7 +1,6 @@
-package Team;
+package team;
 
-import Player.Player;
-import Team.Team;
+import player.Player;
 import match.Match;
 
 import java.util.ArrayList;
