+++ src/test/java/commands/TeamStatComparerTest.java
@@ -1,14 +1,16 @@
+package commands;
+
 import java.util.ArrayList;
 
-import Team.Team;
+import team.Team;
 import commands.TeamStatComparer;
 import commands.TeamStatManager;
 import org.junit.*;
 import static org.junit.Assert.*;
-import Player.Player;
+import player.Player;
 import match.Match;
-import Team.HockeyTeam;
-import Team.TeamManager;
+import team.HockeyTeam;
+import team.TeamManager;
 
 
 public class TeamStatComparerTest {
