+++ src/test/java/commands/TennisPlayerStatManagerTest.java
@@ -1,3 +1,5 @@
+package commands;
+
 import commands.TennisPlayerStatManager;
 import org.junit.*;
