+++ src/test/java/player/TennisPlayerListTest.java
@@ -1,4 +1,6 @@
-import Player.TennisPlayerList;
+package player;
+
+import player.TennisPlayerList;
 import org.junit.*;
 import static org.junit.Assert.*;
