+++ src/test/java/team/TeamManagerTest.java
@@ -1,11 +1,13 @@
+package team;
+
 import java.util.ArrayList;
 
-import Team.Team;
+import team.Team;
 import match.Match;
 import org.junit.*;
-import Player.Player;
-import Team.HockeyTeam;
-import Team.TeamManager;
+import player.Player;
+import team.HockeyTeam;
+import team.TeamManager;
 
 import static org.junit.Assert.*;
